// schema.d.ts
export interface paths {
// ...
    "/3/discover/movie": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Movie
         * @description Find movies using over 30 filters and sort options.
         */
        get: operations["discover-movie"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/movie/{movie_id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Details
         * @description Get the top level details of a movie by ID.
         */
        get: operations["movie-details"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/tv/{series_id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Details
         * @description Get the details of a TV show.
         */
        get: operations["tv-series-details"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/discover/tv": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * TV
         * @description Find TV shows using over 30 filters and sort options.
         */
        get: operations["discover-tv"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/movie/{movie_id}/images": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Images
         * @description Get the images that belong to a movie.
         */
        get: operations["movie-images"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/tv/{series_id}/images": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Images
         * @description Get the images that belong to a TV series.
         */
        get: operations["tv-series-images"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/trending/all/{time_window}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * All
         * @description Get the trending movies, TV shows and people.
         */
        get: operations["trending-all"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/trending/movie/{time_window}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Movies
         * @description Get the trending movies on TMDB.
         */
        get: operations["trending-movies"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/3/trending/tv/{time_window}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * TV
         * @description Get the trending TV shows on TMDB.
         */
        get: operations["trending-tv"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/movie/{movie_id}/credits": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Credits */
        get: operations["movie-credits"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/movie/popular": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Popular
         * @description Get a list of movies ordered by popularity.
         */
        get: operations["movie-popular-list"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/movie/upcoming": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Upcoming
         * @description Get a list of movies that are being released soon.
         */
        get: operations["movie-upcoming-list"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/tv/popular": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Popular
         * @description Get a list of TV shows ordered by popularity.
         */
        get: operations["tv-series-popular-list"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/tv/{series_id}/aggregate_credits": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Aggregate Credits
         * @description Get the aggregate credits (cast and crew) that have been added to a TV show.
         */
        get: operations["tv-series-aggregate-credits"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/tv/{series_id}/similar": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Similar
         * @description Get the similar TV shows.
         */
        get: operations["tv-series-similar"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/tv/{series_id}/season/{season_number}/aggregate_credits": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Aggregate Credits
         * @description Get the aggregate credits (cast and crew) that have been added to a TV season.
         */
        get: operations["tv-season-aggregate-credits"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/3/list/{list_id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Details */
        get: operations["list-details"];
        put?: never;
        post?: never;
        /**
         * Delete
         * @description Delete a list.
         */
        delete: operations["list-delete"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: never;
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "discover-movie": {
        parameters: {
            query?: {
                /** @description use in conjunction with `region` */
                certification?: string;
                /** @description use in conjunction with `region` */
                "certification.gte"?: string;
                /** @description use in conjunction with `region` */
                "certification.lte"?: string;
                /** @description use in conjunction with the `certification`, `certification.gte` and `certification.lte` filters */
                certification_country?: string;
                include_adult?: boolean;
                include_video?: boolean;
                language?: string;
                page?: number;
                primary_release_year?: number;
                "primary_release_date.gte"?: string;
                "primary_release_date.lte"?: string;
                region?: string;
                "release_date.gte"?: string;
                "release_date.lte"?: string;
                sort_by?: "original_title.asc" | "original_title.desc" | "popularity.asc" | "popularity.desc" | "revenue.asc" | "revenue.desc" | "primary_release_date.asc" | "title.asc" | "title.desc" | "primary_release_date.desc" | "vote_average.asc" | "vote_average.desc" | "vote_count.asc" | "vote_count.desc";
                "vote_average.gte"?: number;
                "vote_average.lte"?: number;
                "vote_count.gte"?: number;
                "vote_count.lte"?: number;
                /** @description use in conjunction with `with_watch_monetization_types ` or `with_watch_providers ` */
                watch_region?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_cast?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_companies?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_crew?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_genres?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_keywords?: string;
                with_origin_country?: string;
                with_original_language?: string;
                /** @description can be a comma (`AND`) or pipe (`OR`) separated query */
                with_people?: string;
                /** @description possible values are: [1, 2, 3, 4, 5, 6] can be a comma (`AND`) or pipe (`OR`) separated query, can be used in conjunction with `region` */
                with_release_type?: number;
                "with_runtime.gte"?: number;
                "with_runtime.lte"?: number;
                /** @description possible values are: [flatrate, free, ads, rent, buy] use in conjunction with `watch_region`, can be a comma (`AND`) or pipe (`OR`) separated query */
                with_watch_monetization_types?: string;
                /** @description use in conjunction with `watch_region`, can be a comma (`AND`) or pipe (`OR`) separated query */
                with_watch_providers?: string;
                without_companies?: string;
                without_genres?: string;
                without_keywords?: string;
                without_watch_providers?: string;
                year?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}